Raise error on infinite loop by invalid schedule
Since it's hard to detect invalid cron string, it's better to just raise
error if the calculating loop runs too long.
But how long the threshold is difficult issue.

The issue can be translated into what is the longest cycle for cron.
It is 4 years because of leap year.
(Note that the combination of day-month and wday is OR)

This fixes r7kamura/chrono#2